+++ src/main/java/org/c4marathon/assignment/bankaccount/controller/MainAccountController.java
@@ -4,7 +4,7 @@ import org.c4marathon.assignment.bankaccount.dto.request.SendToSavingRequestDto;
 import org.c4marathon.assignment.bankaccount.dto.response.MainAccountResponseDto;
 import org.c4marathon.assignment.bankaccount.service.MainAccountService;
 import org.c4marathon.assignment.common.annotation.Login;
-import org.c4marathon.assignment.common.session.SessionMemberInfo;
+import org.c4marathon.assignment.member.session.SessionMemberInfo;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
