+++ src/main/java/org/c4marathon/assignment/bankaccount/limit/LimitConst.java
@@ -1,5 +1,8 @@
 package org.c4marathon.assignment.bankaccount.limit;
 
-public class LimitConst {
-	public static int CHARGE_LIMIT = 3000000;
+public final class LimitConst {
+	public static final int CHARGE_LIMIT = 3000000;
+
+	private LimitConst() {
+	}
 }
