+++ src/main/java/org/c4marathon/assignment/bankaccount/product/ProductConst.java
@@ -1,6 +1,8 @@
 package org.c4marathon.assignment.bankaccount.product;
 
-public class ProductConst {
-	public static int DIVIDE_VALUE = 10000; // 이율 계산할 때 나누는 수
+public final class ProductConst {
+	public static final int DIVIDE_VALUE = 10000; // 이율 계산할 때 나누는 수
 
+	private ProductConst() {
+	}
 }
