+++ src/main/java/org/c4marathon/assignment/bankaccount/service/MainAccountServiceImpl.java
@@ -32,7 +32,6 @@ public class MainAccountServiceImpl implements MainAccountService {
 	@Override
 	@Transactional(isolation = Isolation.READ_COMMITTED)
 	public int chargeMoney(long mainAccountPk, int money) {
-		System.out.println("charge start");
 		if (!chargeLimitManager.charge(mainAccountPk, money)) {
 			throw AccountErrorCode.CHARGE_LIMIT_EXCESS.accountException(
 				"MainAccountServiceImpl에서 메인 계좌 충전 중 일일 충전 한도 초과 예외 발생");
@@ -84,9 +83,6 @@ public class MainAccountServiceImpl implements MainAccountService {
 	}
 
 	public boolean isSendValid(int myMoney, int sendMoney) {
-		if (myMoney < sendMoney) {
-			return false;
-		}
-		return true;
+		return myMoney >= sendMoney;
 	}
 }
