+++ src/main/java/org/c4marathon/assignment/common/config/WebConfig.java
@@ -3,7 +3,7 @@ package org.c4marathon.assignment.common.config;
 import java.util.List;
 
 import org.c4marathon.assignment.common.argumentresolver.LoginMemberArgumentResolver;
-import org.c4marathon.assignment.common.interceptor.LoginInterceptor;
+import org.c4marathon.assignment.member.interceptor.LoginInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
