+++ src/main/java/org/c4marathon/assignment/common/exception/CommonExceptionHandler.java
@@ -1,6 +1,7 @@
 package org.c4marathon.assignment.common.exception;
 
 import org.c4marathon.assignment.common.response.ErrorResponse;
+import org.c4marathon.assignment.common.utils.ExceptionLogHelper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -9,8 +10,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 
-import static org.c4marathon.assignment.common.utils.ExceptionLogHelper.*;
-
 @Slf4j
 @RestControllerAdvice
 public class CommonExceptionHandler {
@@ -18,7 +17,7 @@ public class CommonExceptionHandler {
 	@ExceptionHandler(CommonException.class)
 	public ResponseEntity<ErrorResponse> handleCommonException(CommonException error) {
 		ErrorResponse errorResponse = ErrorResponse.of(error.getHttpStatus(), error.getErrorMessage());
-		makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
+		ExceptionLogHelper.makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
 
 		return ResponseEntity.status(error.getHttpStatus()).body(errorResponse);
 	}
@@ -28,7 +27,7 @@ public class CommonExceptionHandler {
 		ErrorCode errorCode = CommonErrorCode.INVALID_ARGUMENT_ERROR;
 		ErrorResponse errorResponse = ErrorResponse.of(errorCode.getHttpStatus(), errorCode.getMessage(),
 			error.getBindingResult());
-		makeExceptionLog(log, error, errorCode.name());
+		ExceptionLogHelper.makeExceptionLog(log, error, errorCode.name());
 
 		return ResponseEntity.status(errorCode.getHttpStatus()).body(errorResponse);
 	}
@@ -37,7 +36,7 @@ public class CommonExceptionHandler {
 	public ResponseEntity<ErrorResponse> handleConstraintViolationException(ConstraintViolationException error) {
 		ErrorCode errorCode = CommonErrorCode.INVALID_ARGUMENT_ERROR;
 		ErrorResponse errorResponse = ErrorResponse.of(errorCode.getHttpStatus(), errorCode.getMessage());
-		makeExceptionLog(log, error, errorCode.name());
+		ExceptionLogHelper.makeExceptionLog(log, error, errorCode.name());
 
 		return ResponseEntity.status(errorCode.getHttpStatus()).body(errorResponse);
 	}
