+++ src/main/java/org/c4marathon/assignment/common/exception/ServerExceptionHandler.java
@@ -1,6 +1,7 @@
 package org.c4marathon.assignment.common.exception;
 
 import org.c4marathon.assignment.common.response.ErrorResponse;
+import org.c4marathon.assignment.common.utils.ExceptionLogHelper;
 import org.c4marathon.assignment.member.exception.MemberException;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
@@ -9,8 +10,6 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import lombok.extern.slf4j.Slf4j;
-
-import static org.c4marathon.assignment.common.utils.ExceptionLogHelper.*;
 
 /**
  * 사용자 정의 예외에 없는 예외를 처리하는 Handler 클래스
@@ -22,7 +21,7 @@ public class ServerExceptionHandler {
 
 	@ExceptionHandler(MemberException.class)
 	public ResponseEntity<ErrorResponse> handleMemberException(MemberException error) {
-		makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
+		ExceptionLogHelper.makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
 		ErrorResponse errorResponse = ErrorResponse.of(error.getHttpStatus(), error.getErrorMessage());
 		return ResponseEntity.status(error.getHttpStatus()).body(errorResponse);
 	}
