+++ src/main/java/org/c4marathon/assignment/member/exception/MemberExceptionHandler.java
@@ -1,13 +1,12 @@
 package org.c4marathon.assignment.member.exception;
 
 import org.c4marathon.assignment.common.response.ErrorResponse;
+import org.c4marathon.assignment.common.utils.ExceptionLogHelper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import lombok.extern.slf4j.Slf4j;
-
-import static org.c4marathon.assignment.common.utils.ExceptionLogHelper.*;
 
 @Slf4j
 @RestControllerAdvice
@@ -15,7 +14,7 @@ public class MemberExceptionHandler {
 
 	@ExceptionHandler(MemberException.class)
 	public ResponseEntity<ErrorResponse> handleMemberException(MemberException error) {
-		makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
+		ExceptionLogHelper.makeExceptionLog(log, error, error.getErrorName(), error.getDebugMessage());
 		ErrorResponse errorResponse = ErrorResponse.of(error.getHttpStatus(), error.getErrorMessage());
 		return ResponseEntity.status(error.getHttpStatus()).body(errorResponse);
 	}
