+++ src/main/java/org/c4marathon/assignment/member/service/MemberService.java
@@ -1,9 +1,9 @@
 package org.c4marathon.assignment.member.service;
 
-import org.c4marathon.assignment.common.session.SessionMemberInfo;
 import org.c4marathon.assignment.member.dto.request.SignInRequestDto;
 import org.c4marathon.assignment.member.dto.request.SignUpRequestDto;
 import org.c4marathon.assignment.member.dto.response.MemberInfoResponseDto;
+import org.c4marathon.assignment.member.session.SessionMemberInfo;
 
 public interface MemberService {
 	void signUp(SignUpRequestDto requestDto);
