+++ src/main/java/org/c4marathon/assignment/member/session/SessionConst.java
@@ -1,5 +1,8 @@
-package org.c4marathon.assignment.common.session;
+package org.c4marathon.assignment.member.session;
 
 public class SessionConst {
 	public static final String MEMBER_INFO = "memberInfo";
+
+	private SessionConst() {
+	}
 }
