+++ src/test/java/org/c4marathon/assignment/bankaccount/controller/MainAccountControllerTest.java
@@ -7,8 +7,8 @@ import org.c4marathon.assignment.bankaccount.dto.response.MainAccountResponseDto
 import org.c4marathon.assignment.bankaccount.limit.LimitConst;
 import org.c4marathon.assignment.bankaccount.service.MainAccountService;
 import org.c4marathon.assignment.common.exception.CommonErrorCode;
-import org.c4marathon.assignment.common.session.SessionConst;
-import org.c4marathon.assignment.common.session.SessionMemberInfo;
+import org.c4marathon.assignment.member.session.SessionConst;
+import org.c4marathon.assignment.member.session.SessionMemberInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
